headlessmc-api/src/main/java/me/earth/headlessmc/api/HasId.java
package me.earth.headlessmc.api;

/**
 * A type which can be identified by an id.
 */
public interface HasId {
    /**
     * @return the id identifying this object.
     */
    int getId();

    static <T extends HasId> T getById(int id, Iterable<T> ids) {
        return getById(String.valueOf(id), ids);
    }

    static <T extends HasId> T getById(String id, Iterable<T> ids) {
        for (T t : ids) {
            if (id.equals(String.valueOf(t.getId()))) {
                return t;
            }
        }

        return null;
    }

}
